Patch example config before running tests with it
The example config is using a geo IP filename that usually does not
exist, because that database is not part of the repository. However,
this causes the config loading to (rightly) fail because of the missing
file. We could simply comment out the MMDB line in the example config,
but where’s the fun in that.

The previous way of doing this was patching the Pydantic Config model
after successfully loading the config, which worked on my machine
because I had the MMDB from the example file available under that name,
but I didn’t notice that it would fail for everyone who didn’t.

Now, we read the YAML and modify the file name before validating the
config using Pydantic.