Run mypy on the package instead of files
This solves a strange import problem I had when writing a custom module
that has the same name as a standard library one. (Will be introduced
with one of the next commits.)

Since the tests are not in any package, and mypy doesn’t allow mixing
`-p` with being given files/directories, we now do two separate mypy
runs.