Couple stderr console & raw stdout
Because even if the console is running on stderr, Rich will still take
print()s to stdout and redirect them to the console, effectively
displaying them on stderr, too.

This change also improves the opt-in handling for commands: Instead of
requesting a raw stdout in the `task_factory`, the setting is now global
for the whole command.

Also, remove a stray `sleep`.