Integrate `UserPhone` & SMS verification code
This was a nasty piece of work, sorry for not committing in smaller
chunks.

I also found out that our database model did not account well for
multiple simultaneous login attempts, or multiple active sign-ins for
that matter, and thus had to redesign it on the fly.

This code still has quite a bunch of TODO markers in it, I’ll tackle
these next.