Simplify login data model & code
Gets rid of the additional sign-in table again by using a third
timestamp. Also removes the blocking table (for now) and instead counts
the incomplete requests. This is good because before, unblocking a User
would have meant to remove the entry in the block table _and_ remove the
_reason_ for blocking them in the first place, else the block would
reappear at the next login attempt.

By neither having the User nor the code be unique, it allows for
multiple simultaneous login attempts (e.g. on desktop & mobile) and even
accounts for the same code being generated twice (it can be used twice
then, too).

Retrying the code generation is therefore no longer necessary either,
which has the nice side effect of no longer needing to do commits or
rollbacks in the query functions, leaving the commit up to the actual
endpoint, which can make sure to only commit once the SMS was sent
successfully.

I’d suggest to revisit this commit once the blocking feature is actually
needed.