Make SMS code timeout configurable
This doesn’t really have a lot to do with what I’m doing in this branch,
but I’ve noticed that it’s not configurable yet despite _appearing_ to
be in the config.

Also, the SMS configuration is moving to another section of the config
and this part would need to move too.