Ensure SQLite is threadsafe
This reverts quite a bit of #95. The idea then was to make sure we're
not using SQLite across threads. The _new_ plan is to _try_ to make sure
we're not using it across threads, but if we are (e.g. because it's
beyond our control), this should be okay, too. So now we're basically
_requiring_ SQLite to support threadsafe operation, and it's no longer
an issue if we perform cross-thread accesses. Which is why we're
disabling `check_same_thread`.