Fix TC006 (quoting the type argument of cast())
There is a long discussion in
<https://github.com/astral-sh/ruff/issues/14676> about how useful this
rule is, especially the performance vs. syntax highlighting question.
Note that apparently `cast()` is not affected by `from __future__ import
annotations`; adding it won't help.

I'm kind of indifferent as to whether this is a good rule or not,
therefore I'll accept ruff's recommendation for now.

Signed-off-by: Tim Weber <scy@scy.name>