Fix S113 (no `requests` timeout)
It's unfortunate that requests doesn't have an easy way to configure a
global, or session, timeout; see
<https://github.com/psf/requests/issues/2011> for details.

I guess we should move to httpx. Fun fact: We currently depend on _both_
httpx _and_ requests, as top-level dependencies, which makes no sense at
all. I've opened #280 to fix that.

@iameru, please let me know if you think 10 seconds is too much or too
little of a timeout.

Signed-off-by: Tim Weber <scy@scy.name>