Let mypy run over code & tests in one call
This avoids calling it twice. Not quite sure why I didn't initially
configure it that way.

Interestingly, it reported 55 files of code and 8 files of tests before,
but now it claims 69 (not 63) files in total. Not sure what's going on
here, but I'll update mypy before I try to debug this further.

Signed-off-by: Tim Weber <scy@scy.name>