Move top-level .gitignore into server/
As far as I can tell, basically all of these are for the backend anyway.
@t-muehlberger, please let me know if I've stolen some that you use in
the frontend. 🙃

I'm doing this so that the build process of the backend doesn't include
ignored files in the resulting packages. While it does read a
`.gitignore` file, it only reads the first one it finds and doesn't
traverse up.

Signed-off-by: Tim Weber <scy@scy.name>