Use pytest and SQLModel versions that don't break
For pytest, any version ≥ 8 will break on
`test_access_config_without_loading` (because the expected exception is
_not_ being raised). This might be because something about the test
setup or fixture handling changed; however, I could not find anything
that looks related in the 8.0 changelog
<https://docs.pytest.org/en/latest/changelog.html#pytest-8-0-0rc1-2023-12-30>.
Will need to look into this in more detail.

For SQLModel, we got lots of type checking and runtime errors on 0.0.11
or higher. I believe this is related to the additional checks they've
introduced there
<https://sqlmodel.tiangolo.com/release-notes/?h=sa_column#0011>. The bad
thing is that it reads like these checks actually protect you from
combining incompatible stuff in your models, which were silently ignored
before, but didn't have the effect intended by the model creator. In
other words: Our models might not be doing what we think they are, and
we should look into this.

Signed-off-by: Tim Weber <scy@scy.name>