Update some dependencies for Python 3.13 support
Unfortunately, Pydantic < 2 seems to throw warnings when running the
tests. There is no more recent 1.10 version available than the one we
have, so I guess we'll fix this by upgrading to Pydantic 2, but for
that, we'll need FastAPI ≥ 0.100. That doesn't seem so far from our
current 0.95.1, but we're currently limited to SQLModel < 0.0.11 (see
706462b6e6f8b46aacffc762268322f5b679fcbd), and SQLModel doesn't support
Pydantic 2 before 0.0.12.

So, fixing the warnings will have to wait until the SQLModel update.

Signed-off-by: Tim Weber <scy@scy.name>