Raise HTTPException on token errors
Before, the InvalidTokenError would just bubble up and cause HTTP 500 in
the endpoint, which isn't nice.

Also, PyJWT checks that the expiration date is in the future by default
(the `verify_exp` option,
<https://pyjwt.readthedocs.io/en/stable/api.html#jwt.decode>), therefore
we can remove the code block that did the extra check and decide which
message to return in the `raise` block.

Signed-off-by: Tim Weber <scy@scy.name>