Update FastAPI, python-multipart, and mypy
Originally wanted to update just python-multipart, but then it
complained about FastAPI's old way of importing it; thus I've decided to
update FastAPI as well. And mypy also made sense, to reduce some of the
false positives.

FastAPI is now using `examples=[…]` instead of `example=…`.

Got rid of some `type: ignore` markers by debugging the root cause.

Updated the log filter to return bools instead of 0 and 1.

The new `type: ignore` markers in `progress.py` are unfortunate, but
that code is complicated and I don't have the time to thoroughly fix it,
and especially make sure that I didn't break anything, right now.

Signed-off-by: Tim Weber <scy@scy.name>